+++ src/fable-library/Util.ts
@@ -1,55 +1,5 @@
 // tslint:disable:ban-types
 
-export type Nullable<T> = T | null | undefined;
-
-export type Option<T> = T | Some<T> | undefined;
-
-// Using a class here for better compatibility with TS files importing Some
-export class Some<T> {
-  public value: T;
-
-  constructor(value: T) {
-    this.value = value;
-  }
-
-  public toJSON() {
-    return this.value;
-  }
-
-  // Don't add "Some" for consistency with erased options
-  public toString() {
-    return String(this.value);
-  }
-
-  public GetHashCode() {
-    return structuralHash(this.value);
-  }
-
-  public Equals(other: Option<T>): boolean {
-    if (other == null) {
-      return false;
-    } else {
-      return equals(this.value, other instanceof Some ? other.value : other);
-    }
-  }
-
-  public CompareTo(other: Option<T>) {
-    if (other == null) {
-      return 1;
-    } else {
-      return compare(this.value, other instanceof Some ? other.value : other);
-    }
-  }
-}
-
-export function value<T>(x: Option<T>) {
-  if (x == null) {
-    throw new Error("Option has no value");
-  } else {
-    return x instanceof Some ? x.value : x;
-  }
-}
-
 // Don't change, this corresponds to DateTime.Kind enum values in .NET
 export const enum DateKind {
   Unspecified = 0,
